+++ src/rot/scheduler/ecs.ts
@@ -12,7 +12,7 @@ export default class EntityScheduler extends Scheduler<string> {
     add(item: string, repeat: boolean) {
         const e = globals.Game!.ecs.getEntity(item);
         if (e === undefined) { throw new Error(`Unknown entity ${item} in scheduler`); }
-        const speedData = getEffectiveSpeedData(e);
+        const speedData = getEffectiveSpeedData(globals.Game!.entityMap, e);
         const speed = speedData?.speed ?? BASE_SPEED;
 
         this._queue.add(item, 1 / speed);
@@ -23,7 +23,7 @@ export default class EntityScheduler extends Scheduler<string> {
         if (this._current !== null && this._repeat.indexOf(this._current) !== -1) {
             const e = globals.Game!.ecs.getEntity(this._current);
             if (e === undefined) { throw new Error(`Unknown entity ${this._current} in scheduler`); }
-            const speedData = getEffectiveSpeedData(e);
+            const speedData = getEffectiveSpeedData(globals.Game!.entityMap, e);
             const speed = speedData?.speed ?? BASE_SPEED;
             this._queue.add(this._current, 1 / speed);
         }